Implement ObjectiveResource
This includes the creation, which happens in of_gamerules.
I also flipped some backslashes to forward slashes

This does not yet allow loading into cp maps, as of_gamerules
does not inherit CTeamplayRoundBasedRules, only CTeamplayRules.

In TF2, this implementation is seperate for Server & Client,
but I felt it appropriate to merge the two into a network class,
in the spirit of making it easier to work with.